♻️ Create a macro package
