+++ bot/bot.go
@@ -1,9 +1,8 @@
 package bot
 
 import (
-	"github.com/NatoBoram/BlackCompany/filter"
+	"github.com/NatoBoram/BlackCompany/adapter"
 	"github.com/NatoBoram/BlackCompany/log"
-	"github.com/aiseeq/s2l/lib/point"
 	"github.com/aiseeq/s2l/lib/scl"
 	"github.com/aiseeq/s2l/protocol/api"
 )
@@ -15,22 +14,6 @@ type Bot struct {
 	miningInitialized bool
 
 	State BotState
-}
-
-type BotState struct {
-	// CcForExp marks command centers that are reserved for new expansions.
-	CcForExp map[api.UnitTag]point.Point
-
-	// CcForOrbitalCommand marks command centers that are reserved for upgrading
-	// to orbital commands.
-	CcForOrbitalCommand api.UnitTag
-
-	// BuildingForAddOn marks a barracks as reserved for building a reactor or
-	// tech lab.
-	BuildingForAddOn api.UnitTag
-
-	// AttackWaves holds the groups of units that are used for attacking.
-	AttackWaves AttackWaves
 }
 
 // Step is called at every step of the game. This is the main loop of the bot.
@@ -46,8 +29,6 @@ func (b *Bot) Step() {
 	}
 
 	b.ParseData()
-
-	b.ExecuteStrategy(&Standard)
 }
 
 // Observe fetches the current observation from the game.
@@ -92,37 +73,13 @@ func (b *Bot) ParseData() {
 
 	if !b.miningInitialized {
 		townHalls := b.FindTownHalls()
-		resources := b.findResourcesNearTownHalls(townHalls)
+		resources := b.FindResourcesNearTownHalls(townHalls)
 		turrets := b.findTurretsNearResourcesNearTownHalls(resources)
-		b.InitMining(filter.ToPoints(turrets))
+		b.InitMining(adapter.ToPoints(turrets))
 		b.miningInitialized = true
 	} else {
 		b.acknowledgeMiners()
 	}
 
 	b.FindClusters() // Not used yet
-}
-
-func (b *Bot) InitState() {
-	b.initCcForExp()
-}
-
-func (b *Bot) initCcForExp() {
-	if b.State.CcForExp == nil {
-		b.State.CcForExp = make(map[api.UnitTag]point.Point)
-	}
-
-	townHalls := b.FindTownHalls()
-	if townHalls.Empty() {
-		log.Warn("Couldn't initialize CcForExp because there are no town halls.")
-		return
-	}
-
-	expansions := append(b.Locs.MyExps, b.Locs.MyStart)
-	for _, expansion := range expansions {
-		townHall := townHalls.ClosestTo(expansion)
-		if townHall.IsCloserThan(1, expansion) {
-			b.State.CcForExp[townHall.Tag] = expansion
-		}
-	}
 }
