+++ bot/strategy.go
@@ -1,5 +1,14 @@
 package bot
 
+// Strategy is a build order to be executed.
+type Strategy struct {
+	Name  string
+	Steps BuildOrder
+}
+
+// BuildOrder is a list of build steps in a strategy.
+type BuildOrder []*BuildStep
+
 // BuildStep is a step in a build order.
 type BuildStep struct {
 	// Name of the step.
@@ -13,26 +22,4 @@ type BuildStep struct {
 
 	// Next determines if we're ready to advance to the next step.
 	Next func(*Bot) bool
-}
-
-// BuildOrder is a list of build steps in a strategy.
-type BuildOrder []*BuildStep
-
-// Strategy is a set of build steps to be executed.
-type Strategy struct {
-	Name  string
-	Steps BuildOrder
-}
-
-// ExecuteStrategy executes a strategy.
-func (b *Bot) ExecuteStrategy(s *Strategy) {
-	for _, step := range s.Steps {
-		if step.Predicate(b) {
-			step.Execute(b)
-		}
-
-		if !step.Next(b) {
-			break
-		}
-	}
 }
