+++ filter/point.go
@@ -1,0 +1,12 @@
+package filter
+
+import (
+	"github.com/aiseeq/s2l/lib/point"
+	"github.com/aiseeq/s2l/lib/scl"
+)
+
+func IsNotAt(point point.Point) scl.Filter {
+	return func(u *scl.Unit) bool {
+		return u.IsCloserThan(1, point)
+	}
+}
