+++ log/level.go
@@ -1,0 +1,17 @@
+package log
+
+type LogLevel int
+
+const (
+	LogLevelDebug LogLevel = iota
+	LogLevelInfo
+	LogLevelWarn
+	LogLevelDanger
+	LogLevelFatal
+)
+
+var level LogLevel = LogLevelDebug
+
+func SetLogLevel(l LogLevel) {
+	level = l
+}
